server/requirements.txt
numpy==1.26.4
scikit-build==0.18.1
ninja==1.13.0
Cython==3.2.1
onnx==1.14.1
onnxruntime==1.15.1
insightface==0.7.3
opencv-contrib-python-headless==4.11.0.86
pillow==11.3.0
scipy==1.13.1
scikit-image==0.24.0
scikit-learn==1.6.1
albumentations==1.3.1
easydict==1.13
prettytable==3.16.0
tqdm==4.67.1
jax==0.6.2
jaxlib==0.6.2
ml_dtypes==0.5.4
opt_einsum==3.4.0
fastapi==0.100.0
starlette==0.27.0
uvicorn==0.21.1
pydantic==1.10.12
fastapi-security==0.5.0
typing_extensions==4.15.0
typing-inspection==0.4.2
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==2.53.0
python-json-logger==4.0.0
redis==7.2.0
fastapi-cache2==0.2.2
httpx==0.28.1
dnspython==2.8.0
python-dotenv==1.2.1
Deprecated==1.3.1
limits==5.8.0
requests==2.32.5
Jinja2==3.1.6
PyJWT==2.11.0
certifi==2026.1.4
Flask==2.3.0
Flask-Cors==3.0.10
Werkzeug==2.3.7
colorama==0.4.6
humanfriendly==10.0
coloredlogs==15.0.1
pymongo==4.16.0
cryptography==42.0.7
hiredis==2.3.2
pyotp==2.9.0
qrcode==7.4.2
torch==2.3.1
